openspec: schematic tests + UI docs cross-link + archive 5 shipped changes
Closes the doable-without-blockers half of the v6 backlog and
archives every openspec that landed in the parallel work this
session.

UI refactor follow-ups (closed):
- 2 new tests in `tests/test_schematic_widget.py` —
  `test_schematic_pixmap_has_logical_size` (HiDPI guard via
  ``devicePixelRatio() >= 1.0``) and
  `test_schematic_repaints_on_theme_change` (light vs. dark
  pixmap bytes differ).
- README Development section + POSITIONING differential #6
  cross-link `docs/UI.md`. POSITIONING also dropped the stale
  "no API/CLI batch" claim (the CLI now ships 10
  subcommands).

Archived (5 changes, 8 already during the previous audit):
- `add-cli-headless-runner` — phases 1-7 shipped + Phase 8
  `cli.rst`. Phase 6 ``validate`` deferred (gated on the
  validation reference set).
- `add-manufacturing-spec-export` — phases 1-4 shipped + CLI
  bonus subcommand. Phase 5 UI menu + Phase 6 docs deferred.
- `add-circuit-export` — phases 1-4 + 7 shipped (LTspice +
  PSIM + Modelica + L(I) builder + CLI). Phase 5 UI + Phase 6
  HTML cross-link + Phase 8 docs deferred.
- `add-crash-reporting` — phases 1 + 3 + 4 shipped (consent
  state machine, scrubber, opt-in init, analytics
  scaffolding). Phase 2 consent dialog UI + Phase 5 docs
  deferred.
- `ui-refactor-followups` — score table, theme_changed
  signal, 3D viewer animations + camera_changed live-sync,
  schematic DPR + theme tests, docs/UI.md cross-link.
  Visual-regression baseline + README screenshot refresh
  deferred.

`openspec/README.md` rewritten — pending count drops from 12
to 7. The 4 new-topology proposals (flyback, LCL,
interleaved, PSFB) plus 3 genuinely-blocked items
(validation: bench data; cascade B-D: benchmark uplift;
code-signed installers: certificates) are what remain.

Includes pre-existing linter-formatting drift on a handful
of scripts that the tooling normalised in this session
(blank-line tightening + import sorting; no behavioural
changes).

+++ src/pfc_inductor/physics/copper.py
@@ -1,4 +1,5 @@
 """DC and AC copper loss for an inductor winding."""
+
 from __future__ import annotations
 
 import math
@@ -32,9 +33,7 @@ def Rac_ohm(
 ) -> float:
     """AC resistance of the winding at frequency f."""
     if wire.type == "litz" and wire.d_strand_mm and wire.n_strands:
-        Fr = Rac_over_Rdc_litz(
-            wire.d_strand_mm * 1e-3, wire.n_strands, f_Hz, layers, T_C
-        )
+        Fr = Rac_over_Rdc_litz(wire.d_strand_mm * 1e-3, wire.n_strands, f_Hz, layers, T_C)
     elif wire.type == "round" and wire.d_cu_mm:
         Fr = Rac_over_Rdc_round(wire.d_cu_mm * 1e-3, f_Hz, layers, T_C)
     else:
@@ -55,14 +54,14 @@ def estimate_layers(N: int, wire: Wire, Wa_mm2: float) -> int:
     window_side_mm = math.sqrt(Wa_mm2)
     if window_side_mm <= 0:
         return 1
-    layers = max(1, int(math.ceil(N * d / window_side_mm)))
+    layers = max(1, math.ceil(N * d / window_side_mm))
     return layers
 
 
 def window_utilization(N: int, wire: Wire, Wa_mm2: float) -> float:
     """Ku = N * A_iso / Wa. Uses outer (insulated) diameter."""
     d_iso_mm = wire.outer_diameter_mm()
-    A_iso_mm2 = math.pi * (d_iso_mm ** 2) / 4.0
+    A_iso_mm2 = math.pi * (d_iso_mm**2) / 4.0
     return (N * A_iso_mm2) / max(Wa_mm2, 1e-9)
